refactor: extract navigation bar into its own component
